requirements.txt
torch
diffusers
transformers
flask
pillow
accelerate
safetensors
